requirements.txt
ultralytics
matplotlib
numpy
opencv-python
pandas
streamlit
python-dotenv
torch
torchvision
Pillow
tqdm
psutil
plotly
numba
